refactor: extending abstract websocket base class
